+++ src/lib.rs
@@ -239,32 +239,36 @@ pub use compiler::utilities::{
     add_related_info, attach_file_to_diagnostics, chain_diagnostic_messages, compare_diagnostics,
     contains_parse_error, create_compiler_diagnostic, create_detached_diagnostic,
     create_diagnostic_collection, create_diagnostic_for_node,
-    create_diagnostic_for_node_from_message_chain, create_symbol_table, create_text_writer,
-    declaration_name_to_string, ensure_script_kind, entity_name_to_string,
-    get_assignment_declaration_kind, get_binary_operator_precedence, get_check_flags,
+    create_diagnostic_for_node_from_message_chain, create_file_diagnostic, create_symbol_table,
+    create_text_writer, declaration_name_to_string, ensure_script_kind, entity_name_to_string,
+    escape_non_ascii_string, format_string_from_args, get_assignment_declaration_kind,
+    get_binary_operator_precedence, get_check_flags, get_compiler_option_value,
     get_containing_function_or_class_static_block, get_declaration_of_kind,
     get_effective_initializer, get_effective_modifier_flags,
     get_effective_modifier_flags_always_include_jsdoc, get_effective_return_type_node,
     get_effective_type_annotation_node, get_element_or_property_access_argument_expression_or_name,
-    get_emit_flags, get_emit_script_target, get_escaped_text_of_identifier_or_literal,
-    get_expression_associativity, get_expression_precedence, get_first_identifier, get_full_width,
-    get_function_flags, get_jsdoc_comment_ranges, get_jsdoc_comments_and_tags,
-    get_jsdoc_type_parameter_declarations, get_language_variant, get_leftmost_expression,
-    get_literal_text, get_object_flags, get_operator_associativity, get_operator_precedence,
+    get_element_or_property_access_name, get_emit_flags, get_emit_script_target,
+    get_escaped_text_of_identifier_or_literal, get_expression_associativity,
+    get_expression_precedence, get_first_identifier, get_full_width, get_function_flags,
+    get_jsdoc_comment_ranges, get_jsdoc_comments_and_tags, get_jsdoc_type_parameter_declarations,
+    get_language_variant, get_leftmost_expression, get_literal_text, get_locale_specific_message,
+    get_object_flags, get_operator_associativity, get_operator_precedence, get_root_declaration,
     get_source_file_of_node, get_syntactic_modifier_flags, get_text_of_identifier_or_literal,
     get_text_of_node_from_source_text, has_dynamic_name, has_static_modifier,
     has_syntactic_modifier, is_access_expression, is_ambient_module, is_any_import_or_re_export,
-    is_assignment_expression, is_assignment_operator, is_bindable_static_element_access_expression,
-    is_block_or_catch_scoped, is_external_or_common_js_module, is_function_block,
+    is_assignment_expression, is_assignment_operator, is_bindable_static_access_expression,
+    is_bindable_static_element_access_expression, is_block_or_catch_scoped, is_dynamic_name,
+    is_entity_name_expression, is_external_or_common_js_module, is_function_block,
     is_function_expression_or_arrow_function, is_import_call, is_in_js_file, is_jsdoc_type_alias,
     is_keyword, is_logical_or_coalescing_assignment_operator, is_object_literal_method,
-    is_property_name_literal, is_string_or_numeric_literal_like, is_super_property,
-    is_this_identifier, is_type_alias, is_type_node_kind, is_write_only_access, modifier_to_flag,
-    modifiers_to_flags, node_is_missing, node_is_present, object_allocator, parse_pseudo_big_int,
-    position_is_synthesized, pseudo_big_int_to_string, set_parent, set_parent_recursive,
-    set_text_range_pos, set_text_range_pos_end, set_text_range_pos_width, set_value_declaration,
-    using_single_line_string_writer, Associativity, FunctionFlags, GetLiteralTextFlags,
-    OperatorPrecedence,
+    is_property_access_entity_name_expression, is_property_name_literal, is_prototype_access,
+    is_string_or_numeric_literal_like, is_super_property, is_this_identifier, is_type_alias,
+    is_type_node_kind, is_variable_like, is_write_only_access, modifier_to_flag,
+    modifiers_to_flags, node_is_missing, node_is_present, node_is_synthesized, object_allocator,
+    parse_pseudo_big_int, position_is_synthesized, pseudo_big_int_to_string, set_parent,
+    set_parent_recursive, set_text_range_pos, set_text_range_pos_end, set_text_range_pos_width,
+    set_value_declaration, using_single_line_string_writer, walk_up_parenthesized_expressions,
+    Associativity, FunctionFlags, GetLiteralTextFlags, OperatorPrecedence,
 };
 use compiler::utilities::{has_invalid_escape, set_localized_diagnostic_messages};
 pub use compiler::utilities_public::{
